refactor(types): modularize event.rs, unify types, use u64 for Transfer.amount
- Split event.rs into transfer.rs and registration.rs modules
- Merge core-types crate into setu-types (single source of truth)
- Change Transfer.amount from i128 to u64 (non-negative amounts)
- Unify SubnetType in subnet.rs with Organization/Personal variants
- Change Unregistration.node_type from String to NodeType enum
- Add StateProvider trait to storage crate (avoid circular deps)
- Update all dependent crates with new imports and builder patterns

Breaking changes:
- Transfer.amount is now u64 (was i128)
- Unregistration.node_type is now NodeType enum (was String)
- core-types crate removed (use setu-types instead)

+++ crates/setu-enclave/src/mock/mod.rs
@@ -205,7 +205,7 @@ impl MockEnclave {
                 resolved_coin.object_id.clone(),
                 &transfer.from,
                 &transfer.to,
-                Some(transfer.amount),
+                Some(transfer.amount as u64),
                 &ctx,
             ).map_err(|e| format!("Runtime error: {}", e))?
         } else {
@@ -222,7 +222,7 @@ impl MockEnclave {
             runtime.execute_simple_transfer(
                 &transfer.from,
                 &transfer.to,
-                transfer.amount,
+                transfer.amount as u64,
                 &ctx,
             ).map_err(|e| format!("Runtime error: {}", e))?
         };
